added comment about volume and real paths
Because the path inside the docker container to the main folder that sonarr downloads to can differ from the real path of this volume, this needs to be corrected when checking whether the subtitle is available. We use a regex for this, but this requires the user to fill in the volume path inside the container and the real path.